refactor: components/common/ 제거
