+++ src/pages/LoginPage/components/FormErrorMessage.tsx
@@ -1,4 +1,4 @@
-import { ErrorMessage } from "@/components/common/Input/FormErrorMessage";
+import { ErrorMessage } from "@/components/Input/FormErrorMessage";
 
 function FormErrorMessage({ errorMessage }: { errorMessage: string }) {
   return <ErrorMessage>{errorMessage}</ErrorMessage>;
